python_ai_service/requirements.txt
# @author Padmin D. Curtis (AI Partner OS3.0) for Fabio Cherici
# @version 1.0.0 (FlorenceEGI — La Bottega)
# @date 2026-04-15
# @purpose Python AI service dependencies — Maestro di Bottega

fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic>=0.52.0
python-dotenv>=1.1.0
pydantic>=2.11.1
